requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
PyPDF2
google-generativeai
